fix: local android login broken (NOTE - first time glitched app first load on local android and web not resolved)
NOTE - in this push notifications PR we first started noticing app loads on Android and Web glitching out with the select organization page popping non stop … unsure which change affected this.

It seems related to the local run of Supabase/Powersync, as switching to PROD mode stops it. Furthermore, I undid all my Flutter changes here and switched back to main and still observed the behavior.

Issue could be with local supabase and powersync running older versions than on remote - though I did update the supabase cli and noticed the issue afterwards …

When I switched to this app config setup with the updated paths for localhost on android, the issue suddenly stopped, and on web too … So this issue is very random at the moment…